Added new A7 data sets to plots. Added drag calculations to makePlot (currently applied to A6)..
CD of A6 as calculated by makePlot (about 0.63) is close to, but not identical with, what Google Sheets calculated (about 0.62). Some is likely due to the "Pinf" setting. In GSheets, we set Pinf manually, whereas the makePlot code uses the max P recorded as Pinf. In some cases, this Pinf is higher than what GSheets shows b/c towards the top of the wind tunnel, the unobstructed airspeeds are faster, and we usually take Pinf to be the free-stream P in the lower half of the wind tunnel.